+++ EssentialFeed/Feed Cache/Infrastructure/Coredata/CoreDataFeedStore+FeedStore.swift
@@ -1,0 +1,53 @@
+//
+//  CoreDataFeedStore+FeedStore.swift
+//  EssentialFeed
+//
+//  Created by Kumar, Sanjay (623) on 04/07/23.
+//
+
+import CoreData
+
+extension CoreDataFeedStore: FeedStore {
+    
+    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        perform { context in
+            do {
+                try ManagedCache.find(in: context).map(context.delete).map(context.save)
+                completion(.success(()))
+            } catch {
+                completion(.failure(error))
+            }
+        }
+    }
+    
+    public func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        perform { context in
+            do {
+                let managedCache = try ManagedCache.newUniqueInstance(in: context)
+                managedCache.timestamp = timestamp
+                managedCache.feed = ManagedFeedImage.images(for: feed, in: context)
+                try context.save()
+                completion(.success(()))
+            } catch {
+                completion(.failure(error))
+            }
+        }
+    }
+    
+    public func retrieve(completion: @escaping RetrievalCompletion) {
+        perform { context in
+            do {
+                if let cache = try ManagedCache.find(in: context) {
+                    completion(.success(
+                        CachedFeed(
+                            feed: cache.localFeed,
+                            timestamp: cache.timestamp)))
+                } else {
+                    completion(.success(.none))
+                }
+            } catch {
+                completion(.failure(error))
+            }
+        }
+    }
+}
